introduce useStore, separate iconColor and iconBackgroundColor, some cleanup

+++ frontend/packages/cliq_term/lib/src/theme/typography.dart
@@ -2,14 +2,14 @@ import 'package:flutter/cupertino.dart';
 
 class TerminalTypography {
   final String fontFamily;
-  final double fontSize;
+  final int fontSize;
 
   const TerminalTypography({required this.fontFamily, required this.fontSize});
 
   TextStyle toTextStyle() {
     return TextStyle(
       fontFamily: fontFamily,
-      fontSize: fontSize,
+      fontSize: fontSize.toDouble(),
       fontWeight: .w500,
       fontVariations: [
         // TODO: figure out a way to use the normal font weight
